+++ simple_version/bt_node.h
@@ -6,7 +6,7 @@
 
 // defining fixed variables and global variables
 #define NODE_SIZE 400
-#define N (400-sizeof(double))/(sizeof(float)+sizeof(double))
+#define N (400-sizeof(double))/(sizeof(double)+sizeof(double))
 #define OVERFLOW_RECS (400-sizeof(double))/sizeof(double)
 
 // declare global variables and objects 
@@ -18,7 +18,7 @@ double resultPtr = -1;
  * Object representing the non-leaf nodes, containing n keys and n+1 pointers to next node. 
 */
 typedef struct _nonleafnode{
-    float keys[N];
+    double keys[N];
     double ptrs[N + 1];
 } NonLeafNode;
 
@@ -26,7 +26,7 @@ typedef struct _nonleafnode{
  * Object representing the leaf nodes, containing n keys and n+1 pointers. 
 */
 typedef struct _leafnode{
-    float keys[N];
+    double keys[N];
     double ptrs[N];
     struct _leafnode *next;
 } LeafNode;
@@ -43,7 +43,7 @@ typedef struct _overflownode{
  * Object containing information of node to be inserted/updated during insertion. 
 */
 typedef struct _insertnode{
-    float key; 
+    double key; 
     double ptr; 
 }InsertNode; 
 
@@ -52,10 +52,10 @@ typedef struct _insertnode{
  * object containing information of nodes to be deleted/updated during deletion. 
 */
 typedef struct _deletenode{
-    float key;
+    double key;
     double ptr; 
-    float oldKey;
-    float newKey; 
+    double oldKey;
+    double newKey; 
 }DeleteNode; 
 
 typedef struct _updateNode{
@@ -66,22 +66,22 @@ typedef struct _updateNode{
 }UpdateNode;
 
 // declare functions
-NonLeafNode* createNonleafNode(double ptr1, float key2, double ptr2);
+NonLeafNode* createNonleafNode(double ptr1, double key2, double ptr2);
 int insertNonLeafNodeKey(NonLeafNode* node, InsertNode* insertNode);
-void insertNLNodeKey(NonLeafNode* node,float key, double ptr,int isLowerbound);
-int searchNonLeafNodeKey(NonLeafNode* node, float key);
-double searchNonLeafNode(NonLeafNode* node, float key);
-int deleteNLNodeKey(NonLeafNode* node, float key);
+void insertNLNodeKey(NonLeafNode* node,double key, double ptr,int isLowerbound);
+int searchNonLeafNodeKey(NonLeafNode* node, double key);
+double searchNonLeafNode(NonLeafNode* node, double key);
+int deleteNLNodeKey(NonLeafNode* node, double key);
 int deleteNonLeafNodeKey(BTPage *page, double root,UpdateNode* updateInfo ,DeleteNode* deleteNode);
 double deleteNonLeafNode(NonLeafNode* node);
-float retrieveNonLeafLowerboundKey(BTPage* page, double node);
+double retrieveNonLeafLowerboundKey(BTPage* page, double node);
 
-LeafNode* createLeafNode(float key, double ptr);
+LeafNode* createLeafNode(double key, double ptr);
 int insertLeafNodeKey(LeafNode* node, InsertNode* insertNode);
-void insertLNodeKey(LeafNode* node,float key, double ptr);
-int searchLeafNodeKey(LeafNode* node, float key);
-double searchLeafNode(LeafNode* node, float key);
-int deleteLNodeKey(LeafNode* node, float key);
+void insertLNodeKey(LeafNode* node,double key, double ptr);
+int searchLeafNodeKey(LeafNode* node, double key);
+double searchLeafNode(LeafNode* node, double key);
+int deleteLNodeKey(LeafNode* node, double key);
 int deleteLeafNodeKey(double root, UpdateNode* updateInfo, DeleteNode* deleteNode);
 double deleteLeafNode(LeafNode* node);
 
@@ -94,11 +94,11 @@ double deleteOverflowNode(BTPage *page,OverflowNode* node, OverflowNode* delNode
 
 int insertKey(BTPage *page, double nodePtr, InsertNode* insertNode);
 int deleteKey(BTPage *page,double root, double nodePtr,DeleteNode* deleteNode, UpdateNode* nodeInfo);
-int getNodes(BTPage* page, double node,float key,float ptr,UpdateNode* nodeInfo);
-void UpdateLowerboundKeys(double node, float oldKey, float newKey);
+int getNodes(BTPage* page, double node,double key,double ptr,UpdateNode* nodeInfo);
+void UpdateLowerboundKeys(double node, double oldKey, double newKey);
 
-double searchKey(BTPage *page, double node, float key);
-double searchRangeKey(BTPage *page, double node, float key);
+double searchKey(BTPage *page, double node, double key);
+double searchRangeKey(BTPage *page, double node, double key);
 
 
 #endif //BT_NODE_H
